refactor(serialization): Move most `_TensorPath` logic to its own file
