+++ tensorizer/serialization.py
@@ -12,7 +12,6 @@ import functools
 import hashlib
 import io
 import itertools
-import json
 import logging
 import operator
 import os
@@ -53,6 +52,7 @@ import tensorizer._crypt as _crypt
 import tensorizer._crypt_info as _crypt_info
 import tensorizer._linear_partition as _linear_partition
 import tensorizer._syscalls as _syscalls
+import tensorizer._tensor_path as _tensor_path
 import tensorizer.stream_io as stream_io
 import tensorizer.utils as utils
 from tensorizer._crypt._cgroup_cpu_count import (
@@ -61,6 +61,11 @@ from tensorizer._crypt._cgroup_cpu_count import (
 from tensorizer._internal_utils import Chunked as _Chunked
 from tensorizer._internal_utils import _variable_read
 from tensorizer._NumpyTensor import _NumpyTensor
+from tensorizer._tensor_path import (
+    _TensorPath,
+    _TensorPathComponent,
+    _TensorPathRegistry,
+)
 
 __all__ = [
     "TensorSerializer",
@@ -104,10 +109,6 @@ cpu_count: int = _effective_cpu_count()
 class _SupportsBool(typing.Protocol):
     def __bool__(self) -> bool: ...
 
-
-# Tensor paths are made up of strings (for mapping keys)
-# and integers (for array indices)
-_TensorPathComponent: "typing.TypeAlias" = Union[str, int]
 
 # Filter functions take either a single string (for a simple top-level dict key)
 # or a tensor path (a sequence of path components, each a string or integer).
@@ -187,160 +188,6 @@ class TensorHash:
     __slots__ = ("type", "hash")
     type: HashType
     hash: bytes
-
-
-class _TensorPath(tuple):
-    def serialized_(self) -> bytes:
-        if self.is_str_:
-            return self[0].encode("utf-8")
-        else:
-            # application/json-seq format
-            return b"\x1e" + json.dumps(
-                self, indent=None, ensure_ascii=True, separators=(",", ":")
-            ).encode("ascii")
-
-    @property
-    def is_str_(self) -> bool:
-        return len(self) == 1 and isinstance(self[0], str)
-
-    def normalize_(self) -> Union[tuple, str]:
-        return self[0] if self.is_str_ else tuple(self)
-
-    def __str__(self) -> str:
-        return str(self.normalize_())
-
-    def append_(self, other: Union[str, int]) -> "_TensorPath":
-        if not isinstance(other, (str, int)):
-            raise TypeError(f"Invalid key type: {other.__class__.__name__!r}")
-        else:
-            return self.__class__(self + (other,))
-
-    def validate_(self) -> None:
-        if not self:
-            raise ValueError("Invalid empty tensor path")
-        for i in self:
-            if not isinstance(i, (str, int)):
-                raise TypeError(
-                    "Invalid tensor path component type:"
-                    f" {i.__class__.__name__!r}"
-                )
-            if isinstance(i, int) and i < 0:
-                raise ValueError(
-                    f"Invalid negative integer tensor path component: {i}"
-                )
-
-    @staticmethod
-    def _invalid_hook(*_args, **_kwargs):
-        raise TypeError("Invalid deserialized type")
-
-    @classmethod
-    def deserialize_(cls, serialized: typing.ByteString) -> "_TensorPath":
-        if not isinstance(serialized, (bytes, bytearray)):
-            raise TypeError(
-                "Invalid tensor path: expected byte string,"
-                f" got {serialized.__class__.__name__!r}"
-            )
-        if not serialized:
-            ret = cls()
-            ret.validate_()
-            return ret
-        first_byte: int = serialized[0]
-        if first_byte == 0x1E:
-            if (
-                len(serialized) < 3
-                or serialized[1] != 0x5B  # "["
-                or serialized[-1] != 0x5D  # "]"
-            ):
-                # Require the form <RS>[...]
-                raise ValueError("Invalid tensor path: non-array json-seq")
-            if 0x0A in serialized or 0x0D in serialized:
-                raise ValueError("Illegal newline in json-seq")
-            try:
-                deserialized: List[Union[str, int]] = json.loads(
-                    serialized[1:],
-                    object_hook=cls._invalid_hook,
-                    parse_float=cls._invalid_hook,
-                    parse_constant=cls._invalid_hook,
-                )
-            except RecursionError as e:
-                raise ValueError(
-                    "Cannot deserialize tensor path due to excessive nesting"
-                ) from e
-            if not isinstance(deserialized, list):
-                raise TypeError(
-                    "Invalid deserialized type:"
-                    " expected array as top level object"
-                )
-            ret = cls(deserialized)
-            ret.validate_()
-            return ret
-        else:
-            return cls((serialized.decode("utf-8"),))
-
-
-@dataclasses.dataclass
-class _TensorPathRegistry:
-    """
-    Tracks tensor paths used so far, throwing an error on prefix conflicts,
-    and building a prefix tree of layers of the nested structure.
-    """
-
-    __slots__ = "_registered_paths"
-    _registered_paths: dict
-
-    def __init__(self):
-        self._registered_paths = {}
-
-    def _check_compatible_types(self, path: _TensorPath) -> None:
-        branch = self._registered_paths
-        for depth, component in enumerate(path):
-            if not branch:
-                break
-            existing_type = type(next(iter(branch)))
-            current_type = type(component)
-            if existing_type is not current_type:
-                prefix: tuple = path[: depth + 1]
-                raise ValueError(
-                    "Conflicting tensor paths:"
-                    f" {path.normalize_()} has a different key type"
-                    f" ({current_type.__name__!r}) than existing keys at the"
-                    f" prefix {prefix} ({existing_type.__name__!r})"
-                )
-            if component not in branch:
-                break
-            branch = branch[component]
-
-    def register_path(self, path: Union[_TensorPath, str]) -> None:
-        branch: dict = self._registered_paths
-        if isinstance(path, str):
-            path = _TensorPath((path,))
-        if not isinstance(path, _TensorPath):
-            raise TypeError(
-                f"Invalid tensor path type: {path.__class__.__name__!r}"
-            )
-        if not path:
-            raise ValueError("Invalid empty tensor path")
-        self._check_compatible_types(path)
-        for component in path[:-1]:
-            branch = branch.setdefault(component, {})
-            if not isinstance(branch, dict):
-                raise ValueError(f"Conflicting tensor paths: {path}, {branch}")
-        component = path[-1]
-        if component in branch:
-            if isinstance(branch[component], dict):
-                raise ValueError(
-                    f"Conflicting tensor paths: {path.normalize_()} is both"
-                    " a leaf and a prefix of another path"
-                )
-            else:
-                raise ValueError(
-                    "Conflicting tensor paths:"
-                    f" {path.normalize_()} is used multiple times"
-                )
-        branch[component] = path
-
-    def dict(self) -> dict:
-        return self._registered_paths
 
 
 @dataclasses.dataclass(order=True)
@@ -786,7 +633,7 @@ class _TensorHeaderDeserializer:
         # Read the name.
         name_slice, offset = self.read_name(buffer, offset)
         with name_slice:
-            self.name: _TensorPath = _TensorPath.deserialize_(bytes(name_slice))
+            self.name: _TensorPath = _TensorPath.deserialize_(name_slice)
         if self.tensor_type != TensorType.STATE_DICT and not self.name.is_str_:
             raise ValueError(
                 "Cannot deserialize structured tensor paths"
@@ -928,11 +775,11 @@ class _MetadataDeserializer(dict):
     @classmethod
     def from_io(
         cls, reader: io.BufferedIOBase, count: int
-    ) -> Tuple["_MetadataDeserializer", dict, bytes]:
+    ) -> Tuple["_MetadataDeserializer", _TensorPathRegistry, bytes]:
         raw = reader.read(cls._total_len_segment.size)
         total_len: int = cls._total_len_segment.unpack(raw)[0]
         if total_len == 0:
-            return cls(), {}, raw
+            return cls(), _TensorPathRegistry(), raw
         else:
             encoded_metadata: bytes = reader.read(total_len)
             raw += encoded_metadata
@@ -941,14 +788,14 @@ class _MetadataDeserializer(dict):
     @classmethod
     def from_buffer(
         cls, buffer: bytes, count: int
-    ) -> Tuple["_MetadataDeserializer", dict]:
+    ) -> Tuple["_MetadataDeserializer", _TensorPathRegistry]:
         offset = 0
         entries = cls()
         registry = _TensorPathRegistry()
         for i in range(count):
             entry, offset = cls._read_entry(buffer, offset, registry)
             entries[entry.name] = entry
-        return entries, registry.dict()
+        return entries, registry
 
     @classmethod
     def _read_entry(
@@ -957,7 +804,7 @@ class _MetadataDeserializer(dict):
         # Read the name.
         name_slice, offset = cls._read_name(buffer, offset)
         with name_slice:
-            name: _TensorPath = _TensorPath.deserialize_(bytes(name_slice))
+            name: _TensorPath = _TensorPath.deserialize_(name_slice)
             registry.register_path(name)
 
         tensor_type = TensorType(buffer[offset])
@@ -1815,7 +1662,7 @@ class TensorDeserializer(
             # This is a list of offsets into the file where the per-tensor data
             # is stored.
             self._metadata: Dict[_TensorPath, TensorEntry]
-            self._metadata, self._structure, self._metadata_raw = (
+            self._metadata, structure, self._metadata_raw = (
                 _MetadataDeserializer.from_io(
                     self._file, self._file_header.tensor_count
                 )
@@ -1830,23 +1677,9 @@ class TensorDeserializer(
                     for name, entry in self._metadata.items()
                     if filter_func(name.normalize_())
                 }
-                # Remove keys from self._structure that aren't in self._metadata
-                structure = self._structure
-                structure_layers = [(structure, iter(tuple(structure)))]
-                while structure_layers:
-                    layer, layer_keys = structure_layers[-1]
-                    for k in layer_keys:
-                        v = layer[k]
-                        if isinstance(v, _TensorPath):
-                            # If this is a leaf, check if it needs to be pruned
-                            if v not in self._metadata:
-                                del layer[k]
-                        else:
-                            # Otherwise, recurse
-                            structure_layers.append((v, iter(tuple(v))))
-                            break
-                    else:
-                        structure_layers.pop()
+                # Remove keys from structure that aren't in self._metadata
+                structure.filter(self._metadata.__contains__)
+            self._structure: Dict[Union[str, int], Any] = structure.dict()
 
             if not isinstance(num_readers, int):
                 raise TypeError(
@@ -2103,84 +1936,6 @@ class TensorDeserializer(
             )
         return tensor
 
-    @staticmethod
-    def _restructure(
-        flat: Dict[_TensorPath, torch.Tensor], use_dict_proxies: bool = False
-    ) -> Union[dict, list, types.MappingProxyType]:
-        for path in flat.keys():
-            if len(path) < 1:
-                raise ValueError("Invalid empty tensor path key")
-
-        # Start reconstructing everything as nested dictionaries
-        base = {}
-        for path, tensor in flat.items():
-            branch = base
-            for component in path[:-1]:
-                branch = branch.setdefault(component, {})
-                if not isinstance(branch, dict):
-                    # Key path conflicts should be caught at the metadata
-                    # parsing step, so this is just an extra sanity check
-                    raise RuntimeError(f"Key path conflict for key {path}")
-            component = path[-1]
-            if component in branch:
-                raise RuntimeError(f"Key path conflict for key {path}")
-            branch[component] = tensor
-
-        # Assign a type to each layer separately
-        def re_type_layer(
-            untyped_layer: dict,
-        ) -> Union[dict, list, types.MappingProxyType]:
-            if use_dict_proxies:
-                return types.MappingProxyType(untyped_layer)
-            is_list = False
-            for key in untyped_layer:
-                if isinstance(key, int):
-                    is_list = True
-                    if key < 0:
-                        raise ValueError(
-                            "Illegal negative integer tensor path component"
-                        )
-                elif is_list:
-                    raise ValueError(
-                        "Invalid tensor path keys:"
-                        " mixes dict and list on same layer"
-                    )
-            if is_list:
-                # Lists are always ordered by the value of their key indices,
-                # rather than the order in the file.
-                list_layer = list(untyped_layer.items())
-                list_layer.sort(key=operator.itemgetter(0))
-                return [v for _, v in list_layer]
-            else:
-                return untyped_layer
-
-        # Track recursive state with a stack.
-        # Iterators track progress through the keys of each layer.
-        # Direct iterators over dictionaries (rather than just keys)
-        # may not be stable while actively mutating the dictionary
-        # mid-iteration, so the iterators are over a stable copy
-        # of the keys instead.
-        base_iter = iter(tuple(base))
-        layers = [(None, None, base, base_iter)]
-        while layers:
-            last_layer, last_key, layer, key_iterator = layers[-1]
-            for k in key_iterator:
-                next_layer = layer[k]
-                if isinstance(next_layer, dict):
-                    # Recurse
-                    next_iter = iter(tuple(next_layer))
-                    layers.append((layer, k, next_layer, next_iter))
-                    break
-            else:
-                # Update the key in the parent (or base) with the corrected type
-                re_typed = re_type_layer(layer)
-                if last_layer is None:
-                    base = re_typed
-                else:
-                    last_layer[last_key] = re_typed
-                layers.pop()
-        return base
-
     def _load_prefixed(
         self,
         prefix: Iterable[_TensorPathComponent],
@@ -2207,7 +1962,7 @@ class TensorDeserializer(
                 unstructured[data.header.name] = data.parameter
                 del data
 
-        restructured = self._restructure(unstructured, use_dict_proxies)
+        restructured = _tensor_path.restructure(unstructured, use_dict_proxies)
 
         if strip_prefix:
             for i in range(prefix_len):
@@ -2851,9 +2606,7 @@ class TensorDeserializer(
         # Results are stored in self._cache
         # Results are then yielded in order with the keys provided
 
-        keys: List[_TensorPath] = [
-            k if isinstance(k, _TensorPath) else _TensorPath((k,)) for k in keys
-        ]
+        keys: List[_TensorPath] = list(map(_TensorPath.wrap_, keys))
         if len(set(keys)) != len(keys):
             raise ValueError("Keys must not have any duplicates")
 
@@ -4318,12 +4071,7 @@ class TensorSerializer:
         next_pos = self._file.tell()
         if _syscalls.has_fallocate() and self._fd:
             size = sum(
-                len(
-                    t.path.serialized_()
-                    if isinstance(t.path, _TensorPath)
-                    else t.path.encode("utf-8")
-                )
-                for t in tensors
+                len(_TensorPath.wrap_(t.path).serialized_()) for t in tensors
             )
             size += sum(
                 t.tensor.element_size()
@@ -4502,45 +4250,6 @@ class TensorSerializer:
 
         self._bulk_write(all_tensors)
 
-    @staticmethod
-    def _key_value_iterator(obj: Union[typing.Sequence, typing.Mapping]):
-        if isinstance(obj, typing.Mapping):
-            for k in obj.keys():
-                if not isinstance(k, str):
-                    raise TypeError(
-                        "Invalid key type for state_dict: expected str, got"
-                        f" {k.__class__.__name__!r}"
-                    )
-            return iter(obj.items())
-        elif isinstance(obj, typing.Sequence):
-            return enumerate(obj)
-        else:
-            raise TypeError(
-                "Cannot serialize type as part of a state_dict:"
-                f" {obj.__class__.__name__!r}"
-            )
-
-    @staticmethod
-    def _flatten_structure(
-        obj: Union[List, typing.Mapping], prefix: _TensorPath = _TensorPath()
-    ) -> Iterable[Tuple[_TensorPath, torch.Tensor]]:
-        iters: List[Tuple[_TensorPath, Iterator]] = [
-            (prefix, TensorSerializer._key_value_iterator(obj))
-        ]
-        while iters:
-            pre, it = iters[-1]
-            for name, item in it:
-                path: _TensorPath = pre.append_(name)
-                if isinstance(item, torch.Tensor):
-                    yield path, item
-                else:
-                    iters.append(
-                        (path, TensorSerializer._key_value_iterator(item))
-                    )
-                    break
-            else:
-                iters.pop()
-
     def write_state_dict(self, state_dict: Union[Dict, List, Tuple]):
         """
         Write the state_dict to the file in Tensorizer format.
@@ -4677,7 +4386,9 @@ class TensorSerializer:
                 tensor=param,
                 callback=None,
             )
-            for name, param in self._flatten_structure(state_dict)
+            for name, param in _tensor_path.flatten_structure(
+                torch.Tensor, state_dict
+            )
         )
